minor module renaming
